dns/resolv: res_ninit()/res_nclose()
