+++ mk/re.mk
@@ -535,15 +535,11 @@ HAVE_EPOLL   := $(shell [ -f $(SYSROOT)/include/sys/epoll.h ] || \
 			[ -f $(SYSROOT)/include/$(MACHINE)/sys/epoll.h ] \
 			&& echo "1")
 endif
-ifneq ($(OS),openbsd)
-HAVE_LIBRESOLV := $(shell [ -f $(SYSROOT)/include/resolv.h ] && echo "1")
-endif
 
-ifneq ($(HAVE_LIBRESOLV),)
-CFLAGS  += -DHAVE_LIBRESOLV
-ifneq ($(OS),freebsd)
-LIBS    += -lresolv
-endif
+HAVE_RESOLV := $(shell [ -f $(SYSROOT)/include/resolv.h ] && echo "1")
+
+ifneq ($(HAVE_RESOLV),)
+CFLAGS  += -DHAVE_RESOLV
 endif
 ifneq ($(HAVE_SYSLOG),)
 CFLAGS  += -DHAVE_SYSLOG
