+++ src/dns/dns.h
@@ -5,7 +5,7 @@
  */
 
 
-#ifdef HAVE_LIBRESOLV
+#ifdef HAVE_RESOLV
 int get_resolv_dns(char *domain, size_t dsize, struct sa *nsv, uint32_t *n);
 #endif
 #ifdef WIN32
